Updating package name in the new tests custom-providers module.
