+++ tests/base/src/test/java/org/keycloak/tests/model/UserConsentModelTest.java
@@ -39,7 +39,7 @@ import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
 import org.keycloak.testframework.remote.annotations.TestOnServer;
 import org.keycloak.testframework.remote.runonserver.InjectRunOnServer;
 import org.keycloak.testframework.remote.runonserver.RunOnServerClient;
-import org.keycloak.testsuite.federation.HardcodedClientStorageProviderFactory;
+import org.keycloak.tests.providers.federation.HardcodedClientStorageProviderFactory;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
