+++ tests/base/src/test/java/org/keycloak/tests/model/UserConsentWithUserStorageModelTest.java
@@ -40,8 +40,8 @@ import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
 import org.keycloak.testframework.remote.annotations.TestOnServer;
 import org.keycloak.testframework.remote.runonserver.InjectRunOnServer;
 import org.keycloak.testframework.remote.runonserver.RunOnServerClient;
-import org.keycloak.testsuite.federation.HardcodedClientStorageProviderFactory;
-import org.keycloak.testsuite.federation.UserMapStorageFactory;
+import org.keycloak.tests.providers.federation.HardcodedClientStorageProviderFactory;
+import org.keycloak.tests.providers.federation.UserMapStorageFactory;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
