+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/components/TestComponentProviderFactory.java
@@ -15,30 +15,12 @@
  * limitations under the License.
  */
 
-package org.keycloak.testsuite.components;
+package org.keycloak.tests.providers.components;
 
-import java.util.List;
-import java.util.Map;
-
-import org.keycloak.provider.Provider;
+import org.keycloak.component.ComponentFactory;
 
 /**
  * @author <a href="mailto:sthorger@redhat.com">Stian Thorgersen</a>
  */
-public interface TestComponentProvider extends Provider {
-
-    DetailsRepresentation getDetails();
-
-    class DetailsRepresentation {
-        Map<String, List<String>> config;
-
-        public Map<String, List<String>> getConfig() {
-            return config;
-        }
-
-        public void setConfig(Map<String, List<String>> config) {
-            this.config = config;
-        }
-    }
-
+public interface TestComponentProviderFactory<T extends TestComponentProvider> extends ComponentFactory<T, TestComponentProvider> {
 }
