+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/federation/DummyUserFederationProvider.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.testsuite.federation;
+package org.keycloak.tests.providers.federation;
 
 import java.util.Arrays;
 import java.util.HashSet;
