+++ tests/custom-providers/src/main/java/org/keycloak/tests/providers/federation/UserMapStorageFactory.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.testsuite.federation;
+package org.keycloak.tests.providers.federation;
 
 import java.util.ArrayList;
 import java.util.List;
