+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -1,1 +1,1 @@
-org.keycloak.testsuite.forms.ClickThroughAuthenticator
+org.keycloak.tests.providers.forms.ClickThroughAuthenticator
