+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.authorization.policy.provider.PolicyProviderFactory
@@ -1,1 +1,1 @@
-org.keycloak.testsuite.authorization.policy.GrantPolicyProvider
+org.keycloak.tests.providers.authorization.policy.GrantPolicyProvider
