+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.broker.provider.IdentityProviderFactory
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.keycloak.testsuite.broker.oidc.OverwrittenMappersTestIdentityProviderFactory
+org.keycloak.tests.providers.broker.oidc.OverwrittenMappersTestIdentityProviderFactory
