+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.provider.Spi
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.keycloak.testsuite.components.TestComponentSpi
+org.keycloak.tests.providers.components.TestComponentSpi
