+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.services.clientpolicy.executor.ClientPolicyExecutorProviderFactory
@@ -1,1 +1,1 @@
-org.keycloak.testsuite.client.policies.TrackEventsClientPolicyExecutor
+org.keycloak.tests.providers.client.policies.TrackEventsClientPolicyExecutor
