+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.storage.UserStorageProviderFactory
@@ -1,2 +1,2 @@
-org.keycloak.testsuite.federation.DummyUserFederationProviderFactory
-org.keycloak.testsuite.federation.UserMapStorageFactory
+org.keycloak.tests.providers.federation.DummyUserFederationProviderFactory
+org.keycloak.tests.providers.federation.UserMapStorageFactory
