+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.storage.client.ClientStorageProviderFactory
@@ -1,1 +1,1 @@
-org.keycloak.testsuite.federation.HardcodedClientStorageProviderFactory
+org.keycloak.tests.providers.federation.HardcodedClientStorageProviderFactory
