+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.tests.providers.components.TestComponentProviderFactory
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.keycloak.testsuite.components.TestComponentImplProviderFactory
+org.keycloak.tests.providers.components.TestComponentImplProviderFactory
